refactoring: comprehensive modularity and code organization overhaul
MAJOR MODULARITY FIXES:
* Eliminate function duplication disasters across codebase
  - Remove duplicate I/O functions (outb, outw, inb) from kernel.c/desktop.c to a dedicated header file
  - Remove duplicate string functions (strlen, strncpy) from kernel.c
  - Extract duplicated font arrays from headers to dedicated source files
  - Create proper modular architecture with separation of concerns

* Reorganize kernel.c and desktop.c with proper structure
  - Move all structs, typedefs, and enums to top of files
  - Consolidate scattered variable declarations into organized sections
  - Moved arrays to dedicated sections
  - Establish clear file organization: includes → defines → structs → variables → arrays → functions

MODULARITY:
* Create src/io.h/io.c - Dedicated x86 I/O port fucntions
* Create src/string.c - String function implementations moved from headers
* Create src/font.c - Font data extracted from headers
* Update src/string.h, src/font.h, src/desktop.h - Convert to header files with declarations only

SHELL QOL:
* Fix empty command handling  pressing Enter on empty prompt now gives new line and prompt instead of error
* Added 'help' command showing all available commands with descriptions

COMPILER WARNING ELIMINATION:
* Fix array bounds warning in ACPI RSDP code with proper pragma directives
* Suppress unused function warnings with __attribute__((unused))
* Fix unused parameter warnings with explicit (void)parameter casting
* Resolve pointer-to-integer casting warnings with proper uintptr_t conversion

CODE QUALITY IMPROVEMENTS:
* Standardize comment style throughout codebase (// format)
* Expand oneline functions into proper multi lineformat for readability
* Add function declarations and forward declarations

TECHNICAL DEBT ELIMINATION:
* Remove modularity disasters where single files handled multiple responsibilities
* Fix poor separation of concerns with dedicated modules for I/O, strings, fonts
* Eliminate scattered variable declarations, structs array etc
* Replace inline implementations in headers with proper source/header separation
