+++ src/font.h
@@ -1,89 +1,8 @@
-static const uint8_t font8x8[128][8] = {
-    [0 ... 31] = {0},
-    [32] = {0x00,0x00,0x00,0x00,0x00,0x00,0x00,0x00},
-    [33] = {0x18,0x3C,0x3C,0x18,0x18,0x00,0x18,0x00},
-    [34] = {0x36,0x36,0x24,0x00,0x00,0x00,0x00,0x00},
-    [35] = {0x36,0x36,0x7F,0x36,0x7F,0x36,0x36,0x00},
-    [36] = {0x0C,0x3E,0x03,0x1E,0x30,0x1F,0x0C,0x00},
-    [37] = {0x00,0x63,0x33,0x18,0x0C,0x66,0x63,0x00},
-    [38] = {0x1C,0x36,0x1C,0x6E,0x3B,0x33,0x6E,0x00},
-    [39] = {0x06,0x06,0x0C,0x00,0x00,0x00,0x00,0x00},
-    [40] = {0x18,0x0C,0x06,0x06,0x06,0x0C,0x18,0x00},
-    [41] = {0x06,0x0C,0x18,0x18,0x18,0x0C,0x06,0x00},
-    [42] = {0x00,0x66,0x3C,0xFF,0x3C,0x66,0x00,0x00},
-    [43] = {0x00,0x0C,0x0C,0x3F,0x0C,0x0C,0x00,0x00},
-    [44] = {0x00,0x00,0x00,0x00,0x0C,0x0C,0x18,0x00},
-    [45] = {0x00,0x00,0x00,0x3F,0x00,0x00,0x00,0x00},
-    [46] = {0x00,0x00,0x00,0x00,0x0C,0x0C,0x00,0x00},
-    [47] = {0x60,0x30,0x18,0x0C,0x06,0x03,0x01,0x00},
+#ifndef _KERNEL_FONT_H
+#define _KERNEL_FONT_H
 
-    [48] = {0x3E,0x63,0x73,0x7B,0x6F,0x67,0x3E,0x00},
-    [49] = {0x0C,0x0E,0x0F,0x0C,0x0C,0x0C,0x3F,0x00},
-    [50] = {0x1E,0x33,0x30,0x1C,0x06,0x33,0x3F,0x00},
-    [51] = {0x1E,0x33,0x30,0x1C,0x30,0x33,0x1E,0x00},
-    [52] = {0x38,0x3C,0x36,0x33,0x7F,0x30,0x78,0x00},
-    [53] = {0x3F,0x03,0x1F,0x30,0x30,0x33,0x1E,0x00},
-    [54] = {0x1C,0x06,0x03,0x1F,0x33,0x33,0x1E,0x00},
-    [55] = {0x3F,0x33,0x30,0x18,0x0C,0x0C,0x0C,0x00},
-    [56] = {0x1E,0x33,0x33,0x1E,0x33,0x33,0x1E,0x00},
-    [57] = {0x1E,0x33,0x33,0x3E,0x30,0x18,0x0E,0x00},
+#include <stdint.h>
+//font data
+extern const uint8_t font8x8[128][8];
 
-    [65] = {0x0C,0x1E,0x33,0x33,0x3F,0x33,0x33,0x00},
-    [66] = {0x3F,0x66,0x66,0x3E,0x66,0x66,0x3F,0x00},
-    [67] = {0x3C,0x66,0x03,0x03,0x03,0x66,0x3C,0x00},
-    [68] = {0x1F,0x36,0x66,0x66,0x66,0x36,0x1F,0x00},
-    [69] = {0x7F,0x46,0x16,0x1E,0x16,0x46,0x7F,0x00},
-    [70] = {0x7F,0x46,0x16,0x1E,0x16,0x06,0x0F,0x00},
-    [71] = {0x3C,0x66,0x03,0x03,0x73,0x66,0x7C,0x00},
-    [72] = {0x33,0x33,0x33,0x3F,0x33,0x33,0x33,0x00},
-    [73] = {0x1E,0x0C,0x0C,0x0C,0x0C,0x0C,0x1E,0x00},
-    [74] = {0x78,0x30,0x30,0x30,0x33,0x33,0x1E,0x00},
-    [75] = {0x67,0x66,0x36,0x1E,0x36,0x66,0x67,0x00},
-    [76] = {0x0F,0x06,0x06,0x06,0x46,0x66,0x7F,0x00},
-    [77] = {0x63,0x77,0x7F,0x7F,0x6B,0x63,0x63,0x00},
-    [78] = {0x63,0x67,0x6F,0x7B,0x73,0x63,0x63,0x00},
-    [79] = {0x1C,0x36,0x63,0x63,0x63,0x36,0x1C,0x00},
-    [80] = {0x3F,0x66,0x66,0x3E,0x06,0x06,0x0F,0x00},
-    [81] = {0x1E,0x33,0x33,0x33,0x3B,0x1E,0x38,0x00},
-    [82] = {0x3F,0x66,0x66,0x3E,0x36,0x66,0x67,0x00},
-    [83] = {0x1E,0x33,0x07,0x0E,0x38,0x33,0x1E,0x00},
-    [84] = {0x3F,0x2D,0x0C,0x0C,0x0C,0x0C,0x1E,0x00},
-    [85] = {0x33,0x33,0x33,0x33,0x33,0x33,0x3F,0x00},
-    [86] = {0x33,0x33,0x33,0x33,0x33,0x1E,0x0C,0x00},
-    [87] = {0x63,0x63,0x63,0x6B,0x7F,0x77,0x63,0x00},
-    [88] = {0x63,0x63,0x36,0x1C,0x1C,0x36,0x63,0x00},
-    [89] = {0x33,0x33,0x33,0x1E,0x0C,0x0C,0x1E,0x00},
-    [90] = {0x7F,0x63,0x31,0x18,0x4C,0x66,0x7F,0x00},
-
-    [97]  = {0x00,0x00,0x1E,0x30,0x3E,0x33,0x6E,0x00},
-    [98]  = {0x07,0x06,0x06,0x3E,0x66,0x66,0x3B,0x00},
-    [99]  = {0x00,0x00,0x1E,0x33,0x03,0x33,0x1E,0x00},
-    [100] = {0x38,0x30,0x30,0x3E,0x33,0x33,0x6E,0x00},
-    [101] = {0x00,0x00,0x1E,0x33,0x3F,0x03,0x1E,0x00},
-    [102] = {0x1C,0x36,0x06,0x0F,0x06,0x06,0x0F,0x00},
-    [103] = {0x00,0x00,0x6E,0x33,0x33,0x3E,0x30,0x1F},
-    [104] = {0x07,0x06,0x36,0x6E,0x66,0x66,0x67,0x00},
-    [105] = {0x0C,0x00,0x1C,0x0C,0x0C,0x0C,0x1E,0x00},
-    [106] = {0x30,0x00,0x30,0x30,0x30,0x33,0x33,0x1E},
-    [107] = {0x07,0x06,0x66,0x36,0x1E,0x36,0x67,0x00},
-    [108] = {0x1C,0x0C,0x0C,0x0C,0x0C,0x0C,0x1E,0x00},
-    [109] = {0x00,0x00,0x36,0x7F,0x7F,0x6B,0x63,0x00},
-    [110] = {0x00,0x00,0x3E,0x33,0x33,0x33,0x33,0x00},
-    [111] = {0x00,0x00,0x1E,0x33,0x33,0x33,0x1E,0x00},
-    [112] = {0x00,0x00,0x3B,0x66,0x66,0x3E,0x06,0x0F},
-    [113] = {0x00,0x00,0x6E,0x33,0x33,0x3E,0x30,0x78},
-    [114] = {0x00,0x00,0x3B,0x6E,0x66,0x06,0x0F,0x00},
-    [115] = {0x00,0x00,0x3E,0x03,0x1E,0x30,0x1F,0x00},
-    [116] = {0x08,0x0C,0x3E,0x0C,0x0C,0x2C,0x18,0x00},
-    [117] = {0x00,0x00,0x33,0x33,0x33,0x33,0x6E,0x00},
-    [118] = {0x00,0x00,0x33,0x33,0x33,0x1E,0x0C,0x00},
-    [119] = {0x00,0x00,0x63,0x6B,0x7F,0x36,0x36,0x00},
-    [120] = {0x00,0x00,0x63,0x36,0x1C,0x36,0x63,0x00},
-    [121] = {0x00,0x00,0x33,0x33,0x33,0x3E,0x30,0x1F},
-    [122] = {0x00,0x00,0x3F,0x19,0x0C,0x26,0x3F,0x00},
-    [123] = {0x38,0x0C,0x0C,0x07,0x0C,0x0C,0x38,0x00},
-    [124] = {0x0C,0x0C,0x0C,0x00,0x0C,0x0C,0x0C,0x00},
-    [125] = {0x07,0x0C,0x0C,0x38,0x0C,0x0C,0x07,0x00},
-    [126] = {0x6E,0x3B,0x00,0x00,0x00,0x00,0x00,0x00},
-    [127] = {0},
-};
+#endif
